refac: move transaction utils to core
