+++ migra/src/clients/mod.rs
@@ -35,3 +35,32 @@ pub use self::mysql::Client as MysqlClient;
 pub mod sqlite;
 #[cfg(feature = "sqlite")]
 pub use self::sqlite::Client as SqliteClient;
+
+pub fn with_transaction<TrxFnMut, Res>(
+    client: &mut AnyClient,
+    trx_fn: &mut TrxFnMut,
+) -> MigraResult<Res>
+where
+    TrxFnMut: FnMut(&mut AnyClient) -> MigraResult<Res>,
+{
+    client
+        .begin_transaction()
+        .and_then(|_| trx_fn(client))
+        .and_then(|res| client.commit_transaction().and(Ok(res)))
+        .or_else(|err| client.rollback_transaction().and(Err(err)))
+}
+
+pub fn maybe_with_transaction<TrxFnMut, Res>(
+    is_needed: bool,
+    client: &mut AnyClient,
+    trx_fn: &mut TrxFnMut,
+) -> MigraResult<Res>
+where
+    TrxFnMut: FnMut(&mut AnyClient) -> MigraResult<Res>,
+{
+    if is_needed {
+        with_transaction(client, trx_fn)
+    } else {
+        trx_fn(client)
+    }
+}
