+++ migra_cli/src/commands/downgrade.rs
@@ -32,7 +32,7 @@ pub(crate) fn rollback_applied_migrations(
         })
         .collect::<Result<Vec<_>, _>>()?;
 
-    database::maybe_with_transaction(
+    migra::maybe_with_transaction(
         opts.transaction_opts.single_transaction,
         &mut client,
         &mut |mut client| {
@@ -41,7 +41,7 @@ pub(crate) fn rollback_applied_migrations(
                 .try_for_each(|(migration_name, content)| {
                     if all_migrations.contains_name(migration_name) {
                         println!("downgrade {}...", migration_name);
-                        database::maybe_with_transaction(
+                        migra::maybe_with_transaction(
                             !opts.transaction_opts.single_transaction,
                             &mut client,
                             &mut |client| client.run_downgrade_migration(migration_name, &content),
